KAN-41(Refactor): Use Decimal in Create and Response DTOs

+++ erp_back/src/modules/quote/dto/create.rs
@@ -8,6 +8,7 @@
 
 use serde::Deserialize;
 use chrono::NaiveDate;
+use rust_decimal::Decimal;
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -15,7 +16,6 @@ pub struct CreateQuoteDto {
     pub client_id: i32,
     pub status_id: i32,
     pub created_at: NaiveDate,
-
     pub details: Vec<CreateQuoteDetailDto>,
 }
 
@@ -23,7 +23,7 @@ pub struct CreateQuoteDto {
 #[serde(rename_all = "camelCase")]
 pub struct CreateQuoteDetailDto {
     pub product_id: i32,
-    pub unit_cost: f64,
-    pub tax: f64,
+    pub unit_cost: Decimal,
+    pub tax: Decimal,
     pub quantity: i32,
 }
